power: oplus_battery_msm8250: Fix FV configuration when QNOVO is enabled
Currently, for CC mode, user space depends on BATT_PROFILE_VOTER's
vote on FV to select a PDO. When Qnovo is enabled, this vote is removed,
causing failure in selecting a PDO and exiting CC mode with QC4 charger.
Fix this, by making PDO selection independent of BATT_PROFILE_VOTER's
vote when qnovo is enabled.

Change-Id: I95817d3e54e373eefbab5625699665517f461322
Signed-off-by: Sahil Chandna <chandna@codeaurora.org>